requirements.txt
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-tinymce==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-auth==2.40.3
idna==3.10
Markdown==3.8.2
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
sqlparse==0.5.3
urllib3==2.5.0
